+++ HISTORY.rst
@@ -7,6 +7,7 @@ v0.4.0 (2017-XX-XX)
 ...................
 * show length in string validation error
 * fix aliases in config during inheritance #55
+* use unicode ellipsis in ``truncate``
 
 v0.3.0 (2017-06-21)
 ...................